Use PNG instead of XPM for the icons in the GTK UI
Starting from version 2.42.11 the gdk-pixbuf library disables the XPM
loader by default. Because of that Fuse won't be able to show the
toolbar icons in the GTK UI.

This patch converts the icons to PNG and embeds them in the binary
using GResource.

The old ui/gtk/pixmaps.c is no longer needed by the GTK UI but it is
still used to generate ui/xlib/xpixmaps.c so it is moved to that
directory.

(fixes bug #509)